requirements.txt
# ゆっくり動画自動生成ツール - 依存関係

# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.28.0
click>=8.0.0

# AI/LLM
openai>=1.0.0
anthropic>=0.3.0
google-genai>=1.20.0

# Audio/Video processing
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
pillow>=9.0.0
moviepy>=1.0.3

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0
pydub>=0.25.0

# Image generation
requests-toolbelt>=0.10.1

# YouTube API
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.90.0

# Web scraping (for trend analysis)
beautifulsoup4>=4.12.0
selenium>=4.10.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Development tools
python-dotenv>=1.0.0
tqdm>=4.65.0
colorama>=0.4.6

# Logging and monitoring
structlog>=23.1.0
loguru>=0.7.0

# Configuration validation
jsonschema>=4.18.0

# File handling
pathlib
shutil
tempfile
uuid

# Date/time handling
datetime
pytz>=2023.3

# HTTP client with retry
tenacity>=8.2.0

# Progress bars
rich>=13.4.0

# Type hints
typing-extensions>=4.7.0 